portal-dymension-xyz icon and Avail da problems
My old pull request was approved by the team and the ticket I opened for PR was closed, so the mods on discord told me that I should open a pull request again, please help me.
When I opened the previous pull reguest, the team approved it on January 25, but there is no Avail icon in front of my portal name in the portal-dymension-xyz portal, and Celestia still shows it. Can you accept this new pull reguest and edit the icon in front of my name in the portal and the hub_seguencer section in Avail?